+++ src/Nethermind/Nethermind.Flashbots.Test/FlashbotsModuleTests.cs
@@ -42,14 +42,14 @@ public partial class FlashbotsModuleTests
                 TestKeysAndAddress.TestBuilderKey.PublicKey,
                 TestKeysAndAddress.TestValidatorKey.PublicKey,
                 TestKeysAndAddress.TestBuilderAddr,
-                block.Header.GasLimit,
-                block.Header.GasUsed,
+                (long)block.Header.GasLimit,
+                (long)block.Header.GasUsed,
                 new UInt256(132912184722469)
             ),
             new RExecutionPayloadV3(ExecutionPayloadV3.Create(block)),
             expectedPayload.BlobsBundle,
             [],
-            block.Header.GasLimit,
+            (long)block.Header.GasLimit,
             new Hash256("0x0000000000000000000000000000000000000000000000000000000000000042")
         );
